wal: xrow buffer cursor
This structure enables to find a xrow buffer row less than given vclock
and then fetch row by row from the xrow forwards to the last appended one.
A xrow buffer cursor is essential to allow the from memory replication
because of a relay which required to be able to fetch all logged rows,
stored in a wal memory (implemented as xrow buffer), from given position
and then follow all new changes.

Part of #3794