replication: use wal memory buffer to fetch rows
Fetch data from wal in-memory buffer. Wal allows to start a fiber
which creates a xrow buffer cursor with given vclock and then fetches
row from the xrow buffer one by one and calls given callback for each
row. Also the wal relaying fiber send a heartbeat message if all
rows were processed there were no rows written for replication timeout
period.
Relay connects to wal with a replica known vclock and tries to
relay data. In case of outdated vclock (wal could not create a cursor
or fetch new row from the cursor) the relay makes a fallback in
order to read logged data from file and then makes another try
to connect to wal with updated vclock and so waiter.
In file mode a relay already has a data to send to a replica so from
not the relay  has not any duty to send heartbeat messages - it
is done by wal relay fiber while it waits for new transactions
written by wal.

Closes #3794