wal: matrix clock structure
A matrix clock which allows to maintain a set of vclocks and
their components order. The main target is to be able to
build a vclock which contains lsns each one is greater or equal
that n (n-majority) corresponding lsn from a matrix clock.

* For instance if we have vclocks like:
 *  1: {1: 10, 2: 12, 3: 0}
 *  2: {1: 10, 2: 14, 3: 1}
 *  3: {1: 0, 2: 8, 3: 4}
 * Then majority 1 builds {1: 10, 2: 14, 3: 4}
 * whereas majority 2 build {1: 10, 2: 12, 3: 1}
 * and majority 3 build {1: 0, 2: 8, 3: 0}.

The purpose of the matrix clock is to evaluate a vclock
which is already processed by wal consumers like relays
or to obtain a majority vclock to commit journal entries
in case of synchronous replication.

Part of #980, #3794