build: tune OSX environment
Added homebrew installation routine as it was suggested in its
instructions. Added installation of the cmake tool. Added upgrade
of the OSX packages to avoid of fails on already existed packages
with the previous versions.

Added reinstallation of the python2 with force option to be able
to install it to /usr/local/ path to make pip install there too.