txn: process synchronization through txn instead of wal
In case of synchronous replication transaction could be
written to wal but rolled back in case of conflict. So it
is not enough to check that wal was synced while checkpoint
or initial join feeding. To handle this txn engine emits a
zero-rows transaction which is passed through wal and then
committed in the same queue as standard transactions. So
successful txn_sync means that all previous transaction
were committed but not just written.

Part of #980