sql: do not change order of inserted values
Before this patch, if an ephemeral space was used during INSERT or
REPLACE, the inserted values were sorted by the first column,
since this was the first part of the index. This can lead to an
error when using the AUTOINCREMENT feature, since changing the
order of the inserted value can change the value inserted instead
of NULL. To avoid this, the patch makes the rowid of the inserted
row in the ephemeral space the first part of the ephemeral space
index. Currently, if the ephemeral space has a rowid, it is always
the last column of the ephemeral space. So, to make rowid the
first part of the index, a cyclic shift is used. Thus, we will not
change the order of all other columns.

Closes #4256